refactor(markdown): share parser across desktop and mobile
