+++ desktop/package.json
@@ -17,6 +17,7 @@
     "desktop:package": "tauri build"
   },
   "dependencies": {
+    "@future-os/markdown": "*",
     "@future-os/thread-projection": "*",
     "@tauri-apps/api": "^2.11.1",
     "@tauri-apps/plugin-dialog": "^2.7.1",
@@ -27,18 +28,14 @@
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
     "react-i18next": "^17.0.8",
-    "remark-gfm": "^4.0.1",
-    "remark-parse": "^11.0.0",
     "shiki": "^4.3.0",
-    "tailwind-merge": "^3.6.0",
-    "unified": "^11.0.5"
+    "tailwind-merge": "^3.6.0"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^9.1.0",
     "@eslint-react/eslint-plugin": "^5.10.0",
     "@tailwindcss/vite": "^4.3.2",
     "@tauri-apps/cli": "^2.11.4",
-    "@types/mdast": "^4.0.4",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^6.0.3",
