+++ mobile/package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "main": "index.ts",
   "dependencies": {
+    "@future-os/markdown": "*",
     "@future-os/thread-projection": "*",
     "@react-native-async-storage/async-storage": "2.2.0",
     "expo": "~57.0.8",
